+++ lib/solargraph/parser/rubyvm/class_methods.rb
@@ -87,7 +87,7 @@ module Solargraph
         end
 
         def infer_literal_node_type node
-          # NodeMethods.infer_literal_node_type node
+          NodeMethods.infer_literal_node_type node
         end
 
         def version
@@ -123,13 +123,12 @@ module Solargraph
           elsif node.type == :DSTR
             here = Range.from_node(node)
             there = Range.from_node(node.children[1])
-            result.push Range.new(here.start, there.start)
+            result.push Range.new(here.start, there&.start || here.ending)
           end
           node.children.each do |child|
             result.concat string_ranges(child)
           end
           if node.type == :DSTR && node.children.last.nil?
-            # result.push Range.new(result.last.ending, result.last.ending)
             last = node.children[-2]
             unless last.nil?
               rng = Range.from_node(last)