+++ spec/yard_map_spec.rb
@@ -1,4 +1,4 @@
-require 'open3'
+require 'fileutils'
 require 'set'
 require 'tmpdir'
 
@@ -111,8 +111,8 @@ describe Solargraph::YardMap do
   end
 
   it 'maps YAML to Psych' do
-    # @todo This breaks in Ruby 2.7
-    next if RUBY_VERSION =~ /^2\.7\./
+    # @todo This breaks in Ruby 2.7+
+    next if RUBY_VERSION =~ /^2\.7\./ || RUBY_VERSION =~ /^3\./
     yard_map = Solargraph::YardMap.new(required: ['yaml'])
     yaml = yard_map.path_pin('YAML')
     expect(yaml.return_type.to_s).to eq('Module<Psych>')
@@ -125,9 +125,12 @@ describe Solargraph::YardMap do
   end
 
   it 'adds automatically imported gems to YardMap' do
-    yard_map = Solargraph::YardMap.new
-    yard_map.change(['bundler/require'].to_set, 'spec/fixtures/workspace-with-gemfile', Set.new)
-    pin = yard_map.path_pin('Backport')
-    expect(pin).to be
+    Dir.mktmpdir do |tmp|
+      FileUtils.cp_r 'spec/fixtures/workspace-with-gemfile', tmp
+      yard_map = Solargraph::YardMap.new
+      yard_map.change(['bundler/require'].to_set, "#{tmp}/workspace-with-gemfile", Set.new)
+      pin = yard_map.path_pin('Backport')
+      expect(pin).to be
+    end
   end
 end